requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
msgspec==0.18.6
pyzmq==26.2.0
requests==2.32.3
urllib3==2.2.3
zeroapi==0.8.0
